requirements.txt
requests>=2.32
google-genai>=1.0
